LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-cross-origin-fetch.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: URL is not valid or contains user credentials.

Harness Error (TIMEOUT), message = null

TIMEOUT Cross-origin fetches from a frame with storage-access are not credentialed by default Test timed out
NOTRUN Cross-origin HTTP redirects from a frame with storage-access are not credentialed by default
